+++ src/main/java/com/example/amt_demo/controller/ArticleController.java
@@ -11,8 +11,7 @@ package com.example.amt_demo.controller;
 import com.example.amt_demo.model.*;
 import com.example.amt_demo.service.ArticleService;
 import com.example.amt_demo.service.CategoryService;
-import com.example.amt_demo.service.PhotoUploadService;
-import org.springframework.lang.Nullable;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,41 +23,14 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.view.RedirectView;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.concurrent.atomic.AtomicInteger;
-
-
 @RequestMapping(path = "/admin/articles")
 @Controller
 
-// DPE - Du moment que vous avez des services autant cacher la logique dans ces services. Et que le controller appel le service.
+@AllArgsConstructor
 public class ArticleController {
 
-    final private PhotoUploadService photoStorageService;
     final private ArticleService articleService;
     final private CategoryService categoryService;
-
-    // DPE - Vous connaissez lombok ? (@AllArgsConstructor)
-
-    /**
-     *
-     * @param photoStorageService
-     * @param articleService
-     * @param categoryService
-     */
-    public ArticleController(PhotoUploadService photoStorageService, ArticleService articleService, CategoryService categoryService) {
-        this.photoStorageService = photoStorageService;
-        this.articleService = articleService;
-        this.categoryService = categoryService;
-    }
-
-    /**
-     *
-     * @param mp
-     * @return
-     */
 
     @GetMapping(path="", produces = {"application/xml"})
     public String getAllCarpets(ModelMap mp) {
@@ -88,22 +60,7 @@ public class ArticleController {
      */
     @PostMapping("/add/post")
     public RedirectView addCarpet(Article newArticle, @RequestParam(name = "images", required = false) MultipartFile[] images, RedirectAttributes redir) {
-        RedirectView redirectView = checkArticleInputs(newArticle, redir, "/admin/articles/add");
-        if (redirectView != null) return redirectView;
-
-        Article tryFind = articleService.findByName(newArticle.getName());
-        if(tryFind != null){
-            redirectView = new RedirectView("/admin/articles/add",true);
-            redir.addFlashAttribute("msg_already_existing_article",true);
-            return redirectView;
-        }
-        articleService.save(newArticle);
-        this.uploadImages(newArticle, images);
-        articleService.save(newArticle);
-
-        redirectView = new RedirectView("/admin/articles",true);
-        redir.addFlashAttribute("msg_article_added",true);
-        return redirectView;
+       return articleService.addCarpet(newArticle, images, redir);
     }
 
     /**
@@ -114,25 +71,8 @@ public class ArticleController {
      * @return
      */
     @GetMapping("{carpet_id}/photo/delete/{id}")
-    public RedirectView deleteCarpetPhoto(@PathVariable String carpet_id, @PathVariable String id, RedirectAttributes redir) {
-        Optional<Article> carpet = articleService.findById(Integer.parseInt(carpet_id));
-
-        if(carpet.isPresent()) {
-            Article c = carpet.get();
-            String path;
-            for(ArticlePhoto cp : c.getPhotos()){
-                if(cp.getId() == Integer.parseInt(id)){
-                    path = cp.getPath();
-                    c.getPhotos().remove(cp);
-                    photoStorageService.delete(path);
-                    break;
-                }
-            }
-            articleService.save(c);
-        }
-        RedirectView redirectView = new RedirectView("/admin/articles/edit/"+carpet_id,true);
-        redir.addFlashAttribute("msg_photo_deleted",true);
-        return redirectView;
+    public RedirectView deleteCarpetPhoto(@PathVariable Integer carpet_id, @PathVariable Integer id, RedirectAttributes redir) {
+        return articleService.deleteCarpetPhoto(carpet_id, id, redir);
     }
 
     /**
@@ -141,34 +81,19 @@ public class ArticleController {
      * @param toAdd
      */
     private boolean handleQuantity(Integer carpetId, Integer toAdd){
-
-        // DPE - Si tu mets cette logique avec l'optional dans le service tu peux gérer que l'objet existe pas avec des exceptions
-        Optional<Article> optional = articleService.findById(carpetId);
-
-        // DPE - https://dev.to/jpswade/return-early-12o5
-        if(optional.isPresent()){
-            Article article = optional.get();
-            Integer current = article.getQuantity();
-            if(current + toAdd >= 0){
-                article.setQuantity(article.getQuantity()+toAdd);
-                articleService.save(article);
-                return true;
-            }
-        }
-        return false;
+        return articleService.handleQuantity(carpetId, toAdd);
     }
 
     /**
      *
-     * @param mp
      * @param id
      * @param redir
      * @return
      */
     @GetMapping("/quantity/increase/{id}")
     // DPE - les variables pas utilisés
-    public RedirectView increaseQuantity(ModelMap mp, @PathVariable String id, RedirectAttributes redir) {
-        this.handleQuantity(Integer.parseInt(id), 1);
+    public RedirectView increaseQuantity(@PathVariable Integer id, RedirectAttributes redir) {
+        this.handleQuantity(id, 1);
         RedirectView redirectView = new RedirectView("/admin/articles",true);
         redir.addFlashAttribute("msg_article_quantity_increase",true);
         redir.addFlashAttribute("articles", articleService.findAll());
@@ -182,8 +107,8 @@ public class ArticleController {
      * @return
      */
     @GetMapping("/quantity/decrease/{id}")
-    public RedirectView decreaseQuantity(@PathVariable String id, RedirectAttributes redir) {
-        if(this.handleQuantity(Integer.parseInt(id), -1)){
+    public RedirectView decreaseQuantity(@PathVariable Integer id, RedirectAttributes redir) {
+        if(this.handleQuantity(id, -1)){
             redir.addFlashAttribute("msg_article_quantity_decrease",true);
         }
         RedirectView redirectView = new RedirectView("/admin/articles",true);
@@ -200,26 +125,8 @@ public class ArticleController {
      */
     @PostMapping("/edit/post")
     public RedirectView editArticle(Article updated, @RequestParam(name = "images", required = false) MultipartFile[] images, RedirectAttributes redir) {
-        String url = "/admin/articles/edit/"+updated.getId();
-        Article tryFind = articleService.findByName(updated.getName());
-        if(tryFind != null && !tryFind.getId().equals(updated.getId())){
-            RedirectView redirectView = new RedirectView(url,true);
-            redir.addFlashAttribute("msg_already_existing_article",true);
-            return redirectView;
-        }
-        RedirectView redirectView1 = checkArticleInputs(updated, redir, url);
-        if (redirectView1 != null) return redirectView1;
-        Optional<Article> optional = articleService.findById(updated.getId());
-        optional.ifPresent(article -> updated.setPhotos(article.getPhotos()));
-        this.uploadImages(updated, images);
-        articleService.save(updated);
-
-        RedirectView redirectView = new RedirectView("/admin/articles/edit/" + updated.getId(),true);
-        redir.addFlashAttribute("msg_article_edited",true);
-        return redirectView;
+        return articleService.editArticle(updated, images, redir);
     }
-
-
 
     /**
      *
@@ -228,22 +135,8 @@ public class ArticleController {
      * @return
      */
     @GetMapping("/edit/{id}")
-    // DPE - Tu ne peux pas dire que ton paramètre est directement un Long ?
-    public String editArticle(ModelMap mp, @PathVariable String id) {
-        mp.addAttribute("editing", true);
-        mp.addAttribute("post_url", "/admin/articles/edit/post");
-        Optional<Article> optional = articleService.findById(Integer.parseInt(id));
-        optional.ifPresent(article -> mp.addAttribute("article", article));
-
-        // DPE - Ne peux tu pas faire une seul map qui connait si la catégorie est checkée ou pas ?
-        List<Category> checked = categoryService.getCategoriesOfCarpet(Integer.valueOf(id));
-        List<Category> notChecked = categoryService.getAllCategories();
-
-        for(Category cat : checked) {
-            notChecked.remove(cat);
-        }
-        mp.addAttribute("categories_checked", checked);
-        mp.addAttribute("categories_not_checked", notChecked);
+    public String editArticle(ModelMap mp, @PathVariable Integer id) {
+        articleService.editArticleForm(mp, id);
         return "admin/articleForm";
     }
 
@@ -255,49 +148,11 @@ public class ArticleController {
      */
     @GetMapping("/delete/{id}")
     public String deleteCarpet(ModelMap mp, @PathVariable String id) {
-        Optional<Article> optional = articleService.findById(Integer.parseInt(id));
-        if(optional.isPresent()){
-            Article article = optional.get();
-            articleService.delete(article);
-            photoStorageService.deleteFolder("carpet-photos/carpet"+ article.getId());
-            mp.addAttribute("msg_article_deleted", true);
-        }
-        mp.addAttribute("articles", articleService.findAll());
+        articleService.deleteCarpet(mp, id);
         return "admin/articles";
     }
 
-    @Nullable
-    private RedirectView checkArticleInputs(Article updated, RedirectAttributes redir, String redirTo) {
-        if(updated.getName().length() == 0){
-            RedirectView redirectView = new RedirectView(redirTo,true);
-            redir.addFlashAttribute("msg_missing_name",true);
-            return redirectView;
-        }
 
-        if(updated.getPrice() == null || updated.getPrice() < 0) updated.setPrice(0.0);
-        if(updated.getQuantity() == null || updated.getQuantity() < 0) updated.setQuantity(0);
-        return null;
-    }
 
-    /**
-     *  @param article
-     * @param images
-     * @return
-     */
-    private boolean uploadImages(Article article, MultipartFile[] images){
-        AtomicInteger i = new AtomicInteger(1);
-        if(images != null) {
-            Arrays.asList(images).stream().forEach(image -> {
-                String filename = image.getOriginalFilename();
-                if (filename.length() > 4) {
-                    String location = "/carpet" + article.getId() + "/";
-                    if (photoStorageService.save(image, location, filename)) {
-                        article.getPhotos().add(new ArticlePhoto(photoStorageService.getRoot() + location + filename));
-                        i.incrementAndGet();
-                    }
-                }
-            });
-        }
-        return i.get() > 1;
-    }
+
 }
