+++ src/main/java/com/example/amt_demo/service/CustomUserDetails.java
@@ -1,22 +1,18 @@
 package com.example.amt_demo.service;
 
+import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.Collection;
 
+@AllArgsConstructor
 public class CustomUserDetails implements UserDetails {
     public static final String ROLE_PREFIX = "ROLE_";
 
     private int id;
     private String username;
     private Collection<? extends GrantedAuthority> authorities;
-
-    public CustomUserDetails(int id, String username, Collection<? extends GrantedAuthority> authorities) {
-        this.id = id;
-        this.username = username;
-        this.authorities = authorities;
-    }
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
