+++ src/main/java/com/example/amt_demo/service/CustomUserService.java
@@ -7,22 +7,11 @@
  */
 
 package com.example.amt_demo.service;
-
-import com.example.amt_demo.auth.UsernameJwtAuthenticationToken;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.Collection;
 
 @Service
 public class CustomUserService {
-
-
     private CustomUserDetails user;
-
 
     public void setUser(CustomUserDetails user){
         this.user = user;
@@ -31,7 +20,5 @@ public class CustomUserService {
     public CustomUserDetails getUser(){
         return user;
     }
-
-
 
 }
