+++ src/test/java/com/example/amt_demo/controller/CartControllerTests.java
@@ -28,9 +28,6 @@ public class CartControllerTests {
     private MockMvc mvc;
 
     @Mock
-    private CartInfoRepository cartInfoRepository;
-
-    @Mock
     private ArticleRepository articleRepository;
